+++ src/components/expenses/expenses.table.component.tsx
@@ -11,10 +11,10 @@ import {
 } from 'react-icons/bs';
 import { connect } from 'react-redux';
 import { bindActionCreators, Dispatch } from 'redux';
-import { Person } from '../models/person.model';
-import { Expense } from '../models/expense.model';
-import { Actions, ApplicationState } from '../store';
-import '../assets/style/table.css';
+import { Person } from '../../models/person.model';
+import { Expense } from '../../models/expense.model';
+import { Actions, ApplicationState } from '../../store';
+import '../../assets/style/table.css';
 
 interface StateProps {
   people: Person[];
