+++ src/components/people/people.table.component.tsx
@@ -10,10 +10,10 @@ import {
 } from 'react-icons/bs';
 import { connect } from 'react-redux';
 import { bindActionCreators, Dispatch } from 'redux';
-import { Person } from '../models/person.model';
-import '../assets/style/table.css';
-import { ApplicationState, Actions } from '../store';
-import { Expense } from '../models/expense.model';
+import { Person } from '../../models/person.model';
+import '../../assets/style/table.css';
+import { ApplicationState, Actions } from '../../store';
+import { Expense } from '../../models/expense.model';
 
 interface StateProps {
   people: Person[];
